Storing the loop counter as a double when evaluating @for directives.

+++ eval.cpp
@@ -92,7 +92,7 @@ namespace Sass {
     env = &new_env;
     Block* body = f->block();
     Expression* val = 0;
-    for (size_t i = lo;
+    for (double i = lo;
          i < hi;
          (*env)[variable] = new (ctx.mem) Number(low->path(), low->line(), ++i)) {
       val = body->perform(this);
